feat : 테스트용 Rq 클래스 분리
